refactor(web-app): create HealthSegmentedBar component
