+++ packages/cube-frontend-web-app/src/components/HealthSegmentedBar/createTimePoints.ts
@@ -1,0 +1,42 @@
+import { Dayjs, ManipulateType } from 'dayjs'
+
+export type TimePoint = {
+  dateTime: Dayjs
+  /**
+   * Milliseconds.
+   */
+  timestamp: number
+  labels: string[]
+}
+
+export type CreateTimeOptions = {
+  now: Dayjs
+  iteration: number
+  value: number
+  unit: ManipulateType
+  labelFormatters: string[]
+}
+
+export const createTimePoints = (options: CreateTimeOptions): TimePoint[] => {
+  const { now, iteration, value, unit, labelFormatters } = options
+
+  const timePoints: TimePoint[] = Array.from(Array(iteration).keys())
+    .map((_, index) => {
+      const dateTime = now.add((index + 1) * value, unit)
+      return {
+        dateTime,
+        timestamp: dateTime.valueOf(),
+        labels: labelFormatters.map((formatter) => dateTime.format(formatter)),
+      } satisfies TimePoint
+    })
+    .reverse()
+
+  timePoints.push({
+    dateTime: now,
+    timestamp: now.valueOf(),
+    // TODO: Integrate 'NOW' with i18n.
+    labels: ['NOW'],
+  })
+
+  return timePoints
+}
