+++ packages/cube-frontend-web-app/src/pages/home/health/_components/healthAccordion/healthAccordionUtils.ts
@@ -1,3 +1,9 @@
+import {
+  createTimePoints,
+  TimePoint,
+} from '@cube-frontend/web-app/components/HealthSegmentedBar/createTimePoints'
+import { TimeRange } from '@cube-frontend/web-app/components/TimeRangeDropdown/timeRangeDropdownUtils'
+import { Dayjs } from 'dayjs'
 import { mockServices } from './mockHealth'
 
 export type ServiceCategory = {
@@ -24,4 +30,48 @@ export const groupServicesByCategory = (
   })
 
   return Array.from(map.values())
+}
+
+export const timePointFns: Record<TimeRange, (now: Dayjs) => TimePoint[]> = {
+  last30Days: (now) =>
+    createTimePoints({
+      now,
+      iteration: 2,
+      value: -15,
+      unit: 'days',
+      labelFormatters: ['MM/DD', 'HH:mm A'],
+    }),
+  last14Days: (now) =>
+    createTimePoints({
+      now,
+      iteration: 2,
+      value: -7,
+      unit: 'days',
+      labelFormatters: ['MM/DD', 'HH:mm A'],
+    }),
+  last7Days: (now) =>
+    createTimePoints({
+      now,
+      iteration: 2,
+      // Use hours because DayJS only support integer values.
+      value: -3.5 * 24,
+      unit: 'hours',
+      labelFormatters: ['MM/DD', 'HH:mm A'],
+    }),
+  last24Hours: (now) =>
+    createTimePoints({
+      now,
+      iteration: 2,
+      value: -12,
+      unit: 'hours',
+      labelFormatters: ['HH:mm A'],
+    }),
+  lastHour: (now) =>
+    createTimePoints({
+      now,
+      iteration: 2,
+      value: -30,
+      unit: 'minutes',
+      labelFormatters: ['HH:mm A'],
+    }),
 }
