--- packages/cube-frontend-web-app/src/pages/home/health/homeHealthUtils.ts
@@ -1,1 +1,0 @@
-export type HealthStatus = 'ok' | 'ng' | 'blank'
